requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
sentence-transformers
chromadb==0.3.29
tqdm
pymupdf
pillow
pytesseract
llama-cpp-python
openai
requests
python-multipart
pymupdf
typing-extensions
qdrant-client
FlagEmbedding
pdfplumber
openpyxl
